requirements.txt
# Core dependencies
pyyaml>=6.0

# Development dependencies (install with: uv sync --dev)
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1.0
mypy>=1.0
safety>=2.0
bandit>=1.7
types-PyYAML>=6.0
